Fix incomplete states (#110)
# Fix incomplete states

When visualizing a running simulation, we sometimes could have incomplete states. This could occur when the client had the last state of the simulation incomplete because this was the most recent data, and when getting the new states, the simulation could have changed of state and the previous one would stay incomplete in the client.

## Modifications

- Filter lock files when getting all states.
- Always send the last state of the client to update it.