Fix legs with same vehicle (#129)
# Fix legs with same vehicle

## Problem

Some passengers could take the same vehicle multiple times. So in the simulation, in the vehicle route, the trip could appear in multiple stops as a boarding or alighting passenger. 

To build the leg for the visualization, we search trough the stops to get the boarding and alighting stop indexes. But it would always find the first occurrences of the trip in the stops, and the indexes would be those of the first time the passenger took the vehicle.

## Modifications

- Add offline mode in CLI.
- Verify how many times the passenger has taken the vehicle before the current leg to get the stop indexes.