Environment variables (#150)
# Environment variables (#150)

## Modifications

- Rework of the environment file mecanism, more details below.
- Publish Dockerfiles:
  - Add environment file copy in python app build.
- Angular build Dockerfile:
  - Rename Dockerfile.
  - Import environment file that will be used by the pre-build script in `npm run build`.
  - Clean build folders.
- Update README.

## Explaination

We chose to not use the `dotenv` librairy that did not simplify the environment variables setup that much.

A single `default-environment.json` contains the environment variables at the root of the repository. This file is then used by both the angular and python parts of the app. 

In the angular project, a `set-environment.ts` script is called in both `npm start` and `npm run build` to update `angular.json` (for  the client port in `ng serve`) and to copy the `default-environment.json` file into the `public` folder. The copy will be used by the running app to access the environment variables (for now, only the server port is needed).

In the python project, more steps are needed. First, we need to copy the `default-environment.json` file into a folder that will be  saved in the python package. Then we also need to read the environment files : first the copied default environment, now in `multimodalsim_viewer/common/environments`, is loaded and then the program check for an `environment.json` file in the current working directory and override the given variables. And finally we need to send this final environment to the web app. This is done by copying the environment into the `environment.json` file in the static folder.

These steps are done each time a script of the python process using an environment variable is run. When building the python package, the only required step is the copy of the default environment and this is done in the build docker script.