fix visualization import error (#169)
# Fix visualization import error (#169)

When importing a visualization, two new items appeared on the list : one corrupted and a second one with a number on the name (occurs when the name is already used). This is due to the utility function to get the path of a simulation folder that was creating the folder if is does not exist. The http handler for the import then checks if the folder already exists to avoid name collisions and find the created folder.

## Modifications

Add a `should_create` parameter to the utility function that defaults to `False`. The function creates the folder only if this parameter is `True`.