Centralized timer service (#193)
# Centralized timer service (#193)

## Goal

There are currently two timers in the application : one in the animation service and one in the visualization service. The first one tries to match the second one. However, this can result in desynchronization problems (entities teleporting, more obvious at high speed).

We can solve this by using a single timer.

## Modifications 

- Create a timer service that is now the reference for : 
  - The visualization time.
  - Whether the visualization is paused or not.
  - Whether the environment for the current time is loading or not.
  - The visualization speed and direction.
- Update the time for each animation cycle. 
- Use and update the visualization variables of the timer service everywhere (UI and animation).
- Save all variables in the local storage on changes and load on initialisation.